Hoang

+++ src/services/eventLikeService.ts
@@ -1,33 +1,89 @@
-import { LikeStatus } from "@prisma/client";
+import { LikeStatus, Prisma } from "@prisma/client";
 import prisma from "../config/prisma";
-import { EventLike, EventLikeInput } from "../types/types";
+import {
+  EventLike,
+  EventLikeFilters,
+  EventLikeInput,
+} from "../types/eventTypes";
+import { ForbiddenError, NotFoundError } from "../error/apiError";
 
-const getAllLikedEvent = async (userId: string): Promise<EventLike[]> => {
-  const allLikedEvents = await prisma.eventLike.findMany({ where: { userId } });
+//NOTE - The user caching should be utilized in this case, instead of
+//searching for new user everytime
+const getAllLikedEvent = async (
+  userId: string,
+  filters: EventLikeFilters = {},
+  pagination: { page?: number; limit?: number }
+): Promise<{ allLikedEvents: EventLike[]; total: number }> => {
+  const {
+    type = "FOOD",
+    status = "OPEN",
+    venueId,
+    dateRange,
+    likeStatus = "ACCEPTED",
+  } = filters;
+  const { page = 1, limit = 10 } = pagination;
+  const skip = (page - 1) * limit;
 
-  if (!allLikedEvents) {
-    throw new Error(`No event has been liked by user ${userId}.`);
+  const prismaUser = await prisma.user.findUnique({
+    where: { id: userId },
+  });
+
+  if (!prismaUser) {
+    throw new NotFoundError(`No user with id ${userId} found!`);
   }
 
-  return allLikedEvents;
+  const where: Prisma.EventLikeWhereInput = {
+    ...(likeStatus && { likeStatus }),
+    event: {
+      status,
+      type,
+      ...(venueId && { venueId }),
+      ...(dateRange && {
+        date: {
+          lte: new Date(dateRange.start),
+          gte: new Date(dateRange.end),
+        },
+      }),
+    },
+  };
+
+  const [allLikedEvents, total] = await Promise.all([
+    prisma.eventLike.findMany({
+      where: { userId },
+      skip,
+      take: limit,
+      orderBy: {
+        event: { date: "asc" },
+      },
+      include: {
+        event: true,
+      },
+    }),
+    prisma.eventLike.count({ where }),
+  ]);
+
+  return { allLikedEvents, total };
 };
 
 const getSpecificLikedEvent = async (id: string): Promise<EventLike> => {
   const likedEvent = await prisma.eventLike.findUnique({ where: { id } });
 
   if (!likedEvent) {
-    throw new Error(`No event has been liked by user with id ${id}.`);
+    throw new NotFoundError("Event like not found or has been deleted!");
   }
 
   return likedEvent;
 };
 
-const createLikeEvent = async (requestedEventLike: EventLikeInput): Promise<EventLike> => {
+const createLikeEvent = async (
+  requestedEventLike: EventLikeInput
+): Promise<EventLike> => {
   const newLikedEvent = await prisma.eventLike.create({
     data: {
       ...requestedEventLike,
       status: LikeStatus.PENDING,
       createdAt: new Date(),
+      updatedAt: new Date(),
     },
   });
 
@@ -38,14 +94,30 @@ const updateLikedEventStatus = async (
   id: string,
   status: LikeStatus
 ): Promise<EventLike> => {
+  const eventLike = await prisma.eventLike.findUnique({ where: { id } });
+  
+  if (!eventLike) {
+    throw new NotFoundError("Event like not found or has been deleted!");
+  }
+  if (eventLike.status === "ACCEPTED" && status === "ACCEPTED") {
+    throw new ForbiddenError("Reject the current like status before changing!");
+  }
+
   const newUpdatedEventLike = await prisma.eventLike.update({
     where: { id },
-    data: { status },
+    data: { status, updatedAt: new Date() },
   });
+
   return newUpdatedEventLike;
 };
 
 const deleteLikedEvent = async (id: string): Promise<void> => {
+  const eventLike = await prisma.eventLike.findUnique({ where: { id } });
+  
+  if (!eventLike) {
+    throw new NotFoundError("Event like not found or has been deleted!");
+  }
+  
   await prisma.eventLike.delete({ where: { id } });
 };
 
@@ -54,5 +126,5 @@ export default {
   getAllLikedEvent,
   createLikeEvent,
   updateLikedEventStatus,
-  deleteLikedEvent
+  deleteLikedEvent,
 };
